fix(ken-476): Address PR comments - TreeIs hashes framed records [no-changelog]
hash_tree_as_is frames every record: a kind byte, then each field as a
fixed-width length and its raw OS bytes, so file content spelling a
record boundary cannot forge an entry and two non-UTF-8 names stay two
names. Nothing persisted holds a TreeIs hash. hash.rs tests move to a
sibling file under the size cap.
